deal with non-vagrant environments

+++ bootstrap-aws.sh
@@ -153,8 +153,6 @@ ____EOF
     found_vagrant=
     if [[ "$(id -u vagrant 2>/dev/null)" == "1000" ]]; then
       userdel -f vagrant || true
-      found_vagrant=1
-    fi
 
     if ! id -u -n ubuntu; then
       groupadd -g 1000 ubuntu
@@ -163,9 +161,15 @@ ____EOF
 
     chown -R ubuntu:ubuntu ~ubuntu
 
+      found_vagrant=1
+    fi
+
+    if [[ -f /tmp/home/.ssh/authorized_keys ]]; then
     mkdir -p ~ubuntu/.ssh
     cp -a /tmp/home/.ssh/authorized_keys ~ubuntu/.ssh/
     chown -R ubuntu:ubuntu ~ubuntu/.ssh
+    fi
+
     install -d -o ubuntu -g ubuntu /data /data/cache /data/git
 
     if [[ -n "$found_vagrant" ]]; then
